Move using Rigidbody instead of manipulating transform directly
Player movement is somewhat stuttery in the game view (not the scene view), but collisions are at least working.

Rigidbody rotation is locked in all axes as we want the player's rotation to be controlled manually, not by the physics engine.

Player collider is no longer a trigger as triggers are not affected by collisions.

Ground detector has been removed in favour of upcoming raycast approach. For now player gets "stuck" to the ground and can't move.

fixedDeltaTime changed back to deltaTime as that seems to be the recommended approach.