Add "air cushion" below player (WIP)
Air cushion is a rigidbody "dropped" beneath the player every frame. When it
collides with the ground, it exerts an upwards force on the player.

Still needs some fine-tuning; right now it takes some time for the upward force
to overcome gravity, at which point the player shoots upwards way too fast!