Add more raycast points for slide detection
This addresses some weird behaviour caused by the player now being frictionless, where if you jumped onto something that contacted the root of a wing, the player would slide "inwards" and hop up onto it. This was caused by the player's momentum carrying it (without friction to decelerate) into a condition where the fuselage raycasts suddenly detected collisions and the player bobbed up.

With more raycasts, there are no realistic cases where this can happen, so the player will always slide away. As an added bonus, this slide will include the player's current momentum, so it appears quite natural.